Implement Distributed Offline-First AI Sync Mesh
Resolves #28036
