+++ src/server/api/mod.rs
@@ -38,5 +38,3 @@ pub mod cart;
 
 pub mod quotes;
 pub mod inbox;
-
-pub mod sync_mesh;
